% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_agreement.R
\name{agreement}
\alias{agreement}
\title{Calculate agreement among annotators}
\usage{
agreement(annotations, annotators = NULL)
}
\description{
Take a list of annotations, by document, annotator, and annotation value (tag), and calculate agreement among a specified subset of annotators.
}
