% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coincidences.R
\name{coincidences_annotator_pair}
\alias{coincidences_annotator_pair}
\title{Calculate annotation coincides for a pair of annotators}
\usage{
coincidences_annotator_pair(annotations, annotators)
}
\description{
Calculate annotation coincides for a pair of annotators
}
\seealso{
Other coincidences: \code{\link{coincidences_annotators_pairwise}},
  \code{\link{coincidences_pooled}}
}
