% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coincidences.R
\name{coincidences_annotators_pairwise}
\alias{coincidences_annotators_pairwise}
\title{Row-bind together all pairwise combinations of annotators, plus optionally
margins}
\usage{
coincidences_annotators_pairwise(annotations,
  annotators = get_annotators(annotations), diag = TRUE,
  impute_annotators = FALSE)
}
\arguments{
\item{annotations}{data.frame containing annotation data}

\item{annotators}{character vector of annotator names}

\item{diag}{Include the diagonal facets showing agreement within annotator}

\item{impute_annotators}{Impute NA tags for annotators who applied no tags}
}
\description{
Coincidences can also be counted within subsets of annotators. One special
case is within the same annotator, i.e. where annotators apply multiple tags,
which do they tend to apply together.
}
\details{
If an annotator is listed in the annotators argument, but doesn't appear in
the annotations dataset, that annotator, that annotator will be imputed to
exist and implicitly have NA tags. Note this may differ from imputation
policies for a given annotation project. [Q: Is that the default behaviour we
want?]
}
\seealso{
Other coincidences: \code{\link{coincidences_annotator_pair}},
  \code{\link{coincidences_pooled}}
}
