% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coincidences.R
\name{coincidences_pooled}
\alias{coincidences_pooled}
\title{Calculate cooccurence of tags pooled across annotators}
\usage{
coincidences_pooled(annotations, annotators = get_annotators(annotations))
}
\arguments{
\item{annotations}{data.frame containing annotation data}

\item{annotators}{character vector of annotator names}
}
\description{
Take a list of annotations, by document, annotator, and annotation value
(tag), and calculate agreement among a specified set of annotators.
}
\seealso{
Other coincidences: \code{\link{coincidences_annotator_pair}},
  \code{\link{coincidences_annotators_pairwise}}
}
