% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_annotators}
\alias{filter_annotators}
\title{Filter to annotations made by specified annotator(s)}
\usage{
filter_annotators(annotations, annotators = NULL)
}
\arguments{
\item{annotations}{data.frame containing annotations}
}
\description{
\code{annotation} data.frames must contain the following columns:
document_name
annotator
annotation_id (in case of multiple annotations per annotator per document)
annotation_type
annotation_value
}
