% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_doc_ids_from_project_path}
\alias{get_doc_ids_from_project_path}
\alias{get_doc_ids}
\alias{get_annotators}
\alias{get_annotation_types}
\alias{get_tagset_used_unsorted}
\alias{get_tagset_used}
\alias{get_tagset_complete}
\title{Getters for annotation attributes stored in \code{annotations}}
\usage{
get_doc_ids_from_project_path(path)

get_doc_ids(annotations, tagged_only = TRUE)

get_annotators(annotations)

get_annotation_types(annotations)

get_tagset_used_unsorted(annotations)

get_tagset_used(annotations, sort = c("value", "frequency",
  "first_appearance"))

get_tagset_complete(annotations, tagset_planned = NULL, sort = c("value",
  "frequency", "first_appearance"))
}
\arguments{
\item{path}{Path to the WebAnno project directory}

\item{annotations}{data.frame of annotations data}

\item{tagged_only}{Logical to indicate whether to impute docs not in
\code{annotations} because no tags applied}

\item{sort}{Order in which to store tagset}

\item{tagset_planned}{character vector of tags available to annotators}

\item{annotations}{data.frame containing annotation data}

\item{sort}{type of sorting on the tagset

  If \code{sort = "first_appearance"} then unused tags are added to the end,
  in the order that they appear in `\code{tagset_planned}.}
}
\description{
Getters for annotation attributes stored in \code{annotations}

Get list of tags used in annotations

Combine used tagset and canonical tagset, and raise warnings if they differ
}
\section{Functions}{
\itemize{
\item \code{get_doc_ids_from_project_path}: Get types of annotation

\item \code{get_doc_ids}: Get types of annotation

\item \code{get_annotation_types}: Get types of annotation

\item \code{get_tagset_used_unsorted}: Get a character vector of the tags used in \code{annotations}

\item \code{get_tagset_used}: Get a character vector of the tags used in \code{annotations}

\item \code{get_tagset_complete}: Get a character vector of the tags used in \code{annotations}
}}

