% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_tagset_complete}
\alias{get_tagset_complete}
\title{Combine used tagset and canonical tagset, and raise warnings if they differ}
\usage{
get_tagset_complete(annotations, tagset_planned = NULL, sort = c("value",
  "frequency", "first_appearance"))
}
\arguments{
\item{sort}{If \code{sort = "first_appearance"} then unused tags are added to the end,
in the order that they appear in `\code{tagset_planned}.}
}
\description{
Combine used tagset and canonical tagset, and raise warnings if they differ
}
