% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_tagset_used}
\alias{get_tagset_used}
\title{Get list of tags used in annotations}
\usage{
get_tagset_used(annotations, sort = c("value", "frequency",
  "first_appearance"))
}
\arguments{
\item{sort}{Order in which to store tagset}
}
\description{
Get list of tags used in annotations
}
