% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coincidences}
\alias{plot_coincidences}
\title{Wrapper to construct the coincidences data.frame and then call the plot function}
\usage{
plot_coincidences(annotations, annotators = get_annotators(annotations))
}
\arguments{
\item{annotations}{data.frame containing annotations data}

\item{annotators}{list of annotators between whom to count coincidences}
}
\description{
Wrapper to construct the coincidences data.frame and then call the plot function
}
