% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coincidences_annotators_pairwise_old}
\alias{plot_coincidences_annotators_pairwise_old}
\alias{plot_coincidences_annotators_pairwise}
\title{Cross-plots of tags from a pair of annotators, faceted by pairs of annotators}
\usage{
plot_coincidences_annotators_pairwise_old(annotations,
  annotators = get_annotators(annotations), tagset_planned = NULL,
  margins = TRUE, cutoff = 0, sort = c("value", "frequency"))

plot_coincidences_annotators_pairwise(annotations,
  annotators = get_annotators(annotations), tagset_planned = NULL,
  margins = TRUE, cutoff = 0, sort = c("value", "frequency"))
}
\arguments{
\item{annotations}{data.frame containing annotations data}

\item{annotators}{list of annotators between whom to count coincidences}

\item{tagset_planned}{character vector of tag names offered to annotators}

\item{margins}{Whether to add marginal plots}

\item{cutoff}{Threshold number of coincidences per cell below which to leave cell blank}

\item{sort}{Method to order the tagset on plot}
}
\description{
Pairwise coccurrences of topics on the same document, from any annotator,
including multiple topic tags applied by same annotator to same document. A
given tag will be counted (m-1) times, where m is the number of tags applied
by any annotator on a given document, i.e. once for each pair of tags that
the tag appears in. Note that to the extent that there are multiple tags from
the same annotator (on a given document) with the tags applied by different
annotators, this plot does *not* provide a completely clean measure of
agreement between annotators (it conflates cooccurence within annotator).
}
\details{
But I don't see a straightforward how to measure agreement where there are
multiple tags per annotator.
}
\section{Functions}{
\itemize{
\item \code{plot_coincidences_annotators_pairwise}: Alternative version of
the function while I try to work around a bug
}}

