% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_marginal_tag_frequency}
\alias{plot_marginal_tag_frequency}
\title{Plot marginal frequency of tags}
\usage{
plot_marginal_tag_frequency(annotations, tagset_planned = NULL,
  sort = c("value", "frequency"), margins = TRUE,
  labeller = long_labeller)
}
\arguments{
\item{annotations}{data.frame containing annotations}

\item{tagset_planned}{Character vector. Use to include unused tags in plot}

\item{sort}{One of "value" or "frequency"}

\item{margins}{TRUE | FALSE}

\item{labeller}{function to format label text for the tagset axis}
}
\description{
Automatically faceted by annotator, with a marginal facet added optionally by
default.
}
